Wrapper on top of the default solver & custom solver
We still need to be able to do "standard" queries to the solver. However the
custom solver has been hacked with only for the specific case of "best effort"
queries.

This implements a wrapper solver, that allows the code to dynamically switch
between the standard solver and the custom solver.