The new cycle-avoiding strategy does not work
opam wants formulas in conflicts to only contain disjunctions (between
packages at least); however, we would like them to have conjunctions
to express a cycle.

This is possible at in the cudf universe, but not allowed in the opam
universe apparently.

(where it fails: when calling OpamSolver.installable, because of
OpamFormula.set_to_disjunction).