neutralize the user environment
(Joint work with Armaël Guéneau)

I had a bug where running Marracheck would setup a new opamroot,
but then try to use the switch set in my environment by OPAMSWITCH,
which of course does not exist in the new root.

This code only works if the sanitization code runs before OPAM caches
the environment. We sent an upstream PR to fix OPAM to not cache the
environment before .init functions are called.