State: rename Versioned into Repo
The naming difficulties around current_timestamp vs. cover_state come
from our initial idea of speaking of ('a Versioned.t) as 'a, as
transparently as possible (as we do for Serialized.t).

I think that we can clarify our naming and code by being explicit
about the fact that a ('a Version.t) is a different object, called
a "repo".