rename the `opam` git submodule to `vendored-opam`
This is because of an unfortunate interaction with opam (the tool)
when doing "opam install .". An `opam` directory has a special
meaning, as a place that can contains the .opam files of the project…
and in our case the `opam` directory does indeed contain .opam files!
But for marracheck itself, but for the vendored opam libraries...

To avoid this confusion, rename the submodule directory, which name
does not matter as we just care that dune explores it recursively.