prepare_universe: include dependencies of base packages as well
this has better semantics. In particular, the code that later checks
if a switch needs to be rebuild would systematically rebuild the
switch even on a fresh switch, because:

- starting from the prepared universe, a cover elt plan could
  (somehow) include the base package without all its dependencies
- the switch would include the base package and all its dependencies

causing the switch to be rebuilt.

I don't know why a cover elt plan would not include all dependencies
of the base packages (the solver should presumably add them back when
computing the plan), but it seems cleaner to have them from the
beginning instead of relying on the solver to add them back.